py/requirements.txt
kafka-python==1.3.5
requests==2.18.4
requests-oauthlib==0.8.0
jsonschema==2.6.0
